+++ code/system.ui/react.common/src/index.mts
@@ -1,4 +1,5 @@
 export { Pkg } from './index.pkg.mjs';
-export { Icon } from './ui.Icon';
-export { Spinner } from './ui.Spinner';
-export { Center } from './ui.Center';
+
+export { Icon } from './ui/Icon';
+export { Spinner } from './ui/Spinner';
+export { Center } from './ui/Center';
