+++ code/system.ui/react.common/src/test.ui/entry.tsx
@@ -3,10 +3,10 @@ import { Dev } from '.';
 import { Pkg } from '../index.pkg.mjs';
 
 const Imports = {
-  'ui.Icon': () => import('../ui.Icon/dev/Icon.SPEC'),
-  'ui.Spinner': () => import('../ui.Spinner/Spinner.SPEC'),
-  'ui.Center': () => import('../ui.Center/Center.SPEC'),
-  'hook.useSizeObserver': () => import('../test.ui.specs/useSizeObserver.SPEC'),
+  'ui.Icon': () => import('../ui/Icon/dev/Icon.SPEC'),
+  'ui.Spinner': () => import('../ui/Spinner/Spinner.SPEC'),
+  'ui.Center': () => import('../ui/Center/Center.SPEC'),
+  'hook.useSizeObserver': () => import('./specs/useSizeObserver.SPEC'),
 };
 
 (async () => {
