+++ code/system.ui/react.common/src/ui/Icon/Icon.View.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { t, css, Color } from '../common';
+import { t, css, Color } from '../../common';
 
 export type IconProps = { size: number; color?: string };
 export type IconComponent = React.FC<IconProps>;
