+++ code/system.ui/react.common/src/ui/Icon/Icon.tsx
@@ -1,4 +1,4 @@
-import { t } from '../common';
+import { t } from '../../common';
 import { IconView as View } from './Icon.View';
 
 import type { IconComponent } from './Icon.View';
