+++ code/system.ui/react.common/src/ui/Icon/dev/Icon.SPEC.tsx
@@ -1,4 +1,4 @@
-import { Spec } from '../../test.ui';
+import { Spec } from '../../../test.ui';
 import { Icons } from './sample';
 
 export default Spec.describe('Icon', (e) => {
