+++ code/system.ui/react.common/src/ui/Icon/dev/sample.tsx
