+++ code/system.ui/react.common/src/ui/Spinner/Spinner.SPEC.tsx
@@ -1,4 +1,4 @@
-import { Spec } from '../test.ui';
+import { Spec } from '../../test.ui';
 import { Spinner } from '.';
 
 export default Spec.describe('Spinner', (e) => {
